move scrollbar to widget content
