+++ src/sql/parts/dashboard/containers/dashboardWidgetContainer.css
@@ -7,12 +7,3 @@ dashboard-widget-container {
 	height: 100%;
 	width: 100%;
 }
-
-.scroll-container {
-	height: 100%;
-	width: 100%;
-}
-
-.scrollable {
-	position: relative;
-}
